Add files via upload
1. Importing COVID19 dataset and preparing it for the analysis by dropping columns and aggregating rows.
2. Deciding on and calculating a good measure for the analysis.
3. Merging two datasets and finding correlations among the data.
4. Visualizing the analysis results using Seaborn.

In this project, I've worked with the COVID19 dataset, published by John Hopkins University, which consists of the data related to the cumulative number of confirmed cases, per day, in each Country.
Also, there is a another dataset consist of various life factors, scored by the people living in each country around the globe. I merged these two datasets to see if there is any relationship between the spread of the virus in a country and how happy people are, living in that country.